refactor(server): split relations into personal-web and mini-map modules
relations.ts had grown into a grab-bag — suggestion feed, value CRUD, hints,
invite materialization, and two subgraph readers. Pull the graph readers into
focused siblings:

- relations-personal.ts: getPersonalWeb plus the shared SubgraphNode/SubgraphEdge
  vocabulary and nodeColumns.
- relations-mini-map.ts: getProfileMiniMap, importing that vocabulary.

Tests split to match (relations-personal.test.ts, relations-mini-map.test.ts).
The per-file auth/db helpers are hoisted into a shared relation-helpers.ts, so
each test file keeps only its own one-line vi.mock("@supabase/ssr") (hoisted
per-file by Vitest, can't be shared) and imports the rest.

Behavior-preserving: 94 server tests pass unchanged.
